+++ lib/IRGen/GenericRequirement.h
@@ -19,10 +19,11 @@
 #ifndef SWIFT_IRGEN_GENERICREQUIREMENT_H
 #define SWIFT_IRGEN_GENERICREQUIREMENT_H
 
+#include "swift/AST/GenericRequirement.h"
 #include "swift/AST/Type.h"
-#include "llvm/ADT/SmallVector.h"
-#include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/DenseMapInfo.h"
+#include "llvm/ADT/STLExtras.h"
+#include "llvm/ADT/SmallVector.h"
 
 namespace llvm {
 class Value;
@@ -40,12 +41,6 @@ namespace irgen {
 class Address;
 class IRGenFunction;
 class IRGenModule;
-
-/// An abstract generic requirement.
-struct GenericRequirement {
-  CanType TypeParameter;
-  ProtocolDecl *Protocol;
-};
 
 using RequirementCallback =
   llvm::function_ref<void(GenericRequirement requirement)>;
@@ -151,26 +146,5 @@ public:
 
 } // end namespace irgen
 } // end namespace swift
-
-namespace llvm {
-  template <> struct DenseMapInfo<swift::irgen::GenericRequirement> {
-    using GenericRequirement = swift::irgen::GenericRequirement;
-    using CanTypeInfo = llvm::DenseMapInfo<swift::CanType>;
-    static GenericRequirement getEmptyKey() {
-      return { CanTypeInfo::getEmptyKey(), nullptr };
-    }
-    static GenericRequirement getTombstoneKey() {
-      return { CanTypeInfo::getTombstoneKey(), nullptr };
-    }
-    static llvm::hash_code getHashValue(GenericRequirement req) {
-      return hash_combine(CanTypeInfo::getHashValue(req.TypeParameter),
-                          hash_value(req.Protocol));
-    }
-    static bool isEqual(GenericRequirement lhs, GenericRequirement rhs) {
-      return (lhs.TypeParameter == rhs.TypeParameter &&
-              lhs.Protocol == rhs.Protocol);
-    }
-  };
-}
 
 #endif
